Fenetre

+++ serveur1-No11.c
@@ -180,7 +180,7 @@ int main(int argc, char* argv[]){
                                     MAP_SHARED | MAP_ANONYMOUS, -1,0);
     *shared_memory_seq = 1;
 
-    uint8_t *shared_memory_window = mmap(NULL, PAGESIZE,
+    uint16_t *shared_memory_window = mmap(NULL, PAGESIZE,
                                     PROT_READ | PROT_WRITE,
                                     MAP_SHARED | MAP_ANONYMOUS, -1,0);
     *shared_memory_window = 90;
